requirements.txt
pandas
biopython
streamlit
